feat: 장소 리뷰 시스템 및 질문 API

+++ src/main/java/com/dingdong/eeum/dto/response/PlaceDetailResponseDto.java
@@ -24,7 +24,7 @@ public class PlaceDetailResponseDto {
     private String email;
     private String website;
     private double temperature;
-    private int reviewCount;
+    private long reviewCount;
     private PlaceStatus status;
     private boolean isVerified;
